feat: Add metadata field to EventDto
The code changes add a new `metadata` field to the `EventDto` object in the `common.ts` file. This field is of type `z.record(z.any())` and allows for storing additional metadata related to the event.

Based on the recent commits and repository conventions, the suggested commit message is "feat: Add metadata field to EventDto".